+++ packages/effect/test/Tracer.test.ts
@@ -258,4 +258,22 @@ describe("Tracer", () => {
         assert.strictEqual(onEndCalled, true)
       }))
   })
+
+  it.effect("withTracerEnabled", () =>
+    Effect.gen(function*($) {
+      const span = yield* $(
+        Effect.currentSpan,
+        Effect.withSpan("A"),
+        Effect.withTracerEnabled(false)
+      )
+      const spanB = yield* $(
+        Effect.currentSpan,
+        Effect.withSpan("B"),
+        Effect.withTracerEnabled(true)
+      )
+
+      assert.deepEqual(span.name, "A")
+      assert.deepEqual(span.spanId, "noop")
+      assert.deepEqual(spanB.name, "B")
+    }))
 })