+++ packages/opentelemetry/src/Resource.ts
@@ -2,7 +2,14 @@
  * @since 1.0.0
  */
 import * as Resources from "@opentelemetry/resources"
-import { SemanticResourceAttributes, TelemetrySdkLanguageValues } from "@opentelemetry/semantic-conventions"
+import {
+  SEMRESATTRS_SERVICE_NAME,
+  SEMRESATTRS_SERVICE_VERSION,
+  SEMRESATTRS_TELEMETRY_SDK_LANGUAGE,
+  SEMRESATTRS_TELEMETRY_SDK_NAME,
+  TELEMETRYSDKLANGUAGEVALUES_NODEJS,
+  TELEMETRYSDKLANGUAGEVALUES_WEBJS
+} from "@opentelemetry/semantic-conventions"
 import { GenericTag } from "effect/Context"
 import * as Layer from "effect/Layer"
 
@@ -29,16 +36,16 @@ export const layer = (config: {
   readonly serviceVersion?: string
   readonly attributes?: Resources.ResourceAttributes
 }) => {
-  const attributes = {
+  const attributes: Record<string, string> = {
     ...(config.attributes ?? undefined),
-    [SemanticResourceAttributes.SERVICE_NAME]: config.serviceName,
-    [SemanticResourceAttributes.TELEMETRY_SDK_NAME]: "@effect/opentelemetry",
-    [SemanticResourceAttributes.TELEMETRY_SDK_LANGUAGE]: typeof (globalThis as any).document === "undefined"
-      ? TelemetrySdkLanguageValues.NODEJS
-      : TelemetrySdkLanguageValues.WEBJS
+    [SEMRESATTRS_SERVICE_NAME]: config.serviceName,
+    [SEMRESATTRS_TELEMETRY_SDK_NAME]: "@effect/opentelemetry",
+    [SEMRESATTRS_TELEMETRY_SDK_LANGUAGE]: typeof (globalThis as any).document === "undefined"
+      ? TELEMETRYSDKLANGUAGEVALUES_NODEJS
+      : TELEMETRYSDKLANGUAGEVALUES_WEBJS
   }
   if (config.serviceVersion) {
-    attributes[SemanticResourceAttributes.SERVICE_VERSION] = config.serviceVersion
+    attributes[SEMRESATTRS_SERVICE_VERSION] = config.serviceVersion
   }
   return Layer.succeed(
     Resource,