+++ packages/platform-node/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @effect/platform-node
 
+## 0.45.17
+
+### Patch Changes
+
+- [#2384](https://github.com/Effect-TS/effect/pull/2384) [`2f488c4`](https://github.com/Effect-TS/effect/commit/2f488c436de52576562803c57ebc132ef40ccdd8) Thanks [@tim-smart](https://github.com/tim-smart)! - update dependencies
+
+- Updated dependencies [[`2f488c4`](https://github.com/Effect-TS/effect/commit/2f488c436de52576562803c57ebc132ef40ccdd8), [`37ca592`](https://github.com/Effect-TS/effect/commit/37ca592a4101ad90adbf8c8b3f727faf3110cae5), [`317b5b8`](https://github.com/Effect-TS/effect/commit/317b5b8e8c8c2207469b3ebfcf72bf3a9f7cbc60)]:
+  - @effect/platform-node-shared@0.3.15
+  - effect@2.4.11
+  - @effect/platform@0.48.15
+
+## 0.45.16
+
+### Patch Changes
+
+- Updated dependencies [[`9bab1f9`](https://github.com/Effect-TS/effect/commit/9bab1f9fa5b999740755e4e82485cb77c638643a), [`9bbde5b`](https://github.com/Effect-TS/effect/commit/9bbde5be9a0168d1c2a0308bfc27167ed62f3968)]:
+  - effect@2.4.10
+  - @effect/platform@0.48.14
+  - @effect/platform-node-shared@0.3.14
+
 ## 0.45.15
 
 ### Patch Changes